requirements.txt
aiohttp
nextcord
python-dotenv
aiosqlite